+++ src/macos/mod.rs
@@ -93,7 +93,7 @@ impl BatteryReadout for MacOSBatteryReadout {
         )))
     }
 
-    fn health(&self) -> Result<u64, ReadoutError> {
+    fn health(&self) -> Result<u8, ReadoutError> {
         Err(ReadoutError::NotImplemented)
     }
 }
