+++ src/windows/mod.rs
@@ -53,7 +53,7 @@ impl BatteryReadout for WindowsBatteryReadout {
         }
     }
 
-    fn health(&self) -> Result<u64, ReadoutError> {
+    fn health(&self) -> Result<u8, ReadoutError> {
         Err(ReadoutError::NotImplemented)
     }
 }
